requirements.txt
Flask==3.0.0
requests==2.31.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
PyJWT==2.8.0
python-dotenv==1.0.0
cryptography==41.0.6
pytest==7.4.0
flake8==6.0.0
openapi-core==0.19.2
PyYAML==6.0.3
prometheus-client==0.17.1
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-instrumentation-flask==0.43b0
opentelemetry-instrumentation-requests==0.43b0
